requirements.txt
celery==5.4.0
curl_cffi==0.7.1
Django==5.0.8
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
fuzzywuzzy[speedup]==0.18.0
gunicorn==23.0.0
itsdangerous==2.2.0
psycopg2-binary==2.9.9
pytz==2024.1
requests-toolbelt==1.0.0
selectolax==0.3.21
setuptools==75.1.0
beautifulsoup4==4.12.3
djangorestframework_simplejwt==5.4.0
django-filter==23.2
